backend/requirements.txt
pandas
matplotlib
seaborn
azure-functions
azure-storage-blob
